update : 기능 단위로 smoke test 분리(로그인 및 회원정보 수정)
